Removing MatchResourceV3.java
